trying to open server ports

+++ run.sh
@@ -1,9 +1,14 @@
 #! /bin/bash
 set -exo nounset
 
-[ $# -eq 2 ] || exit 1
+[ $# -ge 1 ] || exit 1
+[ -f $1/port -o $# -eq 2 ] || exit 1
 T="`sed 's#/$##' <<< "$1"`"
-P=$2
+if [ -f $1/port ] ; then
+	P="`cat $1/port`"
+else
+	P=$2
+fi
 
 trap "sudo docker rmi -f $T" 0
